gogabangla/requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.7
django-appconf==1.0.2
django-floppyforms==1.7.0
django-model-utils==3.1.2
django-notifications-hq==1.5.0
django-select2==6.3.1
django-select2-forms==2.0.2
django-select2-tags==0.0.1
django-select2light==0.3
django-sortedm2m==1.5.0
django-tastypie==0.14.2
idna==2.7
jsonfield==2.0.2
langdetect==1.0.7
oauthlib==2.1.0
Pillow==5.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-mimeparse==1.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
reverse==0.1.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
